Actualización Reto DRAG N DROP (mapa)
+Se creo script Manager Collector
