+++ jungle/Assets/Art/Sprites/Containers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2bf0785ca5e027b4bb34f4165d6c116c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
