+++ jungle/Assets/Prefabs/Jungle/Containers/trash1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 46b9d17607b73b344bb6b7a09e14f7e3
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
