+++ jungle/Assets/Prefabs/Jungle/Containers/trash2.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e93d2f1becaf7354fb5ed32c11f78a00
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
