--- jungle/Assets/Scripts/Container.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c4d3c337c68d3b94a8bda00cbf543352
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
