+++ jungle/Assets/Scripts/Drag and Drop.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dd8ee7ea81195db4a91d1bd4bf9d3e93
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
