+++ jungle/Assets/Scripts/Drag and Drop/CollectorManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fba89e4b3b7c38948956c5d574c1a16d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
