--- jungle/Assets/Scripts/GarbagePull.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8da7192fce78eb54694e111196be7ead
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
